Work in progress, implement channels in transport and add TODO comment regarding `interact` pattern in signer.
