+++ packages/signer-web/.gitignore
@@ -1,0 +1,12 @@
+node_modules/
+dist/
+lib/
+**/*.js
+**/*.js.map
+**/*.d.ts
+
+# generated docs
+/docs/reference
+
+# Cannot ignore .d.ts files in types/
+!types/**/*.d.ts
