+++ packages/signer-web/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "@slide-computer/signer-web",
+  "version": "1.10.0",
+  "author": "Slide",
+  "license": "MIT",
+  "description": "JavaScript and TypeScript library to communicate with web signers on the Internet Computer",
+  "homepage": "https://github.com/slide-computer/signer-js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/slide-computer/signer-js.git",
+    "directory": "packages/signer-web"
+  },
+  "bugs": {
+    "url": "https://github.com/slide-computer/signer-js/issues"
+  },
+  "keywords": [
+    "internet computer",
+    "internet-computer",
+    "ic",
+    "dfinity",
+    "signer",
+    "wallet",
+    "actor",
+    "dfx",
+    "canister",
+    "candid",
+    "motoko",
+    "javascript",
+    "typescript",
+    "blockchain",
+    "crypto",
+    "distributed",
+    "api",
+    "sdk"
+  ],
+  "main": "./lib/cjs/index",
+  "module": "./lib/esm/index",
+  "unpkg": "./lib/esm/index",
+  "scripts": {
+    "build": "tsc -b && tsc -p tsconfig-cjs.json"
+  },
+  "peerDependencies": {
+    "@dfinity/agent": "^1.0.1",
+    "@dfinity/candid": "^1.0.1",
+    "@dfinity/identity": "^1.0.1",
+    "@dfinity/principal": "^1.0.1",
+    "@slide-computer/signer": "^1.10.0",
+    "@slide-computer/signer-storage": "^1.10.0"
+  },
+  "dependencies": {
+    "buffer": "^6.0.3"
+  },
+  "devDependencies": {
+    "esbuild": "^0.15.16",
+    "eslint": "^8.56.0",
+    "prettier": "^3.2.5",
+    "typescript": "^5.3.3"
+  }
+}
