+++ packages/signer-web/src/icrc29/index.ts
@@ -1,1 +1,2 @@
 export * from "./postMessageTransport";
+export * from "./postMessageChannel";
