+++ packages/signer-web/tsconfig-cjs.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig-cjs.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "composite": true,
+    "outDir": "./lib/cjs",
+    "paths": {
+      "@slide-computer/signer": ["../signer"]
+    },
+    "rootDir": "./src"
+  },
+  "include": [
+    "types/*",
+    "src/**/*"
+  ],
+  "references": [{ "path": "../signer" }]
+}
