+++ packages/signer-web/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "composite": true,
+    "outDir": "./lib/esm",
+    "paths": {
+      "@slide-computer/signer": ["../signer"]
+    },
+    "rootDir": "./src"
+  },
+  "include": [
+    "types/*",
+    "src/**/*"
+  ],
+  "references": [{ "path": "../signer" }]
+}
